refactor: 피드백 반영 리팩토링
