+++ src/main/java/lotto/domain/LottoNumber.java
@@ -1,10 +1,9 @@
-package lotto;
+package lotto.domain;
+
 
 import java.util.Objects;
 
 public class LottoNumber {
-    private final static int LOTTO_START_NUMBER = 1;
-    private final static int LOTTO_END_NUMBER = 45;
     private final static String LOTTO_NUMBER_VALIDATION_MESSAGE = "로또 숫자는 1 ~ 45 사이의 숫자만 가능합니다.";
     private final int lottoNumber;
 
@@ -14,7 +13,7 @@ public class LottoNumber {
     }
 
     private void validation(int number) {
-        if (number < LOTTO_START_NUMBER || number > LOTTO_END_NUMBER) {
+        if (number < LottoMagicNumber.START.get() || number > LottoMagicNumber.END.get()) {
             throw new IllegalArgumentException(LOTTO_NUMBER_VALIDATION_MESSAGE);
         }
     }
