+++ src/test/java/lotto/domain/LottoMachineTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeAll;
@@ -8,6 +8,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
 import static org.assertj.core.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class LottoMachineTest {
     private static LottoMachine lottoMachine;
@@ -15,6 +16,30 @@ public class LottoMachineTest {
     @BeforeAll
     static void beforeAll() {
         lottoMachine = new LottoMachine();
+    }
+
+
+    @DisplayName("구입 할 수 있는 로또의 개수를 구한다")
+    @ParameterizedTest
+    @CsvSource(value = {"10000,10","15000,15","1000,1"})
+    public void buyLottos(int price, int expected) throws Exception {
+        //when
+        int purchaseQuantity = lottoMachine.purchaseQuantity(price);
+
+        //then
+        assertEquals(expected, purchaseQuantity);
+    }
+
+    @Test
+    @DisplayName("1000원 미만의 금액으로 로또를 구매하려고 하면 오류")
+    public void purchasePriceLimit() {
+        //given
+        int price = 900;
+
+        //when then
+        assertThatExceptionOfType(IllegalArgumentException.class)
+                .isThrownBy(() -> lottoMachine.purchaseQuantity(price))
+                .withMessageMatching("1000원 이상의 금액을 입력해야합니다.");
     }
 
     @Test
@@ -32,8 +57,7 @@ public class LottoMachineTest {
     @CsvSource(value = {"10000,10","15000,15","1000,1"})
     public void manyGenerate(int price, int expected) {
         //given
-        Lotto lotto = new Lotto(price);
-        int purchaseQuantity = lotto.purchaseQuantity();
+        int purchaseQuantity = lottoMachine.purchaseQuantity(price);
 
         //when
         LottoTickets lottoTickets = lottoMachine.generate(purchaseQuantity);
